+++ src/cellarr_array/cellarray_dense.py
@@ -7,7 +7,7 @@ from typing import List, Tuple, Union
 
 import numpy as np
 
-from .cellarray import CellArray
+from .cellarray_base import CellArray
 from .helpers import SliceHelper
 
 __author__ = "Jayaram Kancherla"
